# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Geto API",
    "description": "Geto Database Management Tool API",
    "license": {
      "name": ""
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/connections/{id}/history": {
      "get": {
        "tags": [
          "crate::routes::query"
        ],
        "operationId": "list_history_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List query execution history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::query"
        ],
        "operationId": "clear_history_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Clear query history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearHistoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
    "/api/connections/{id}/query": {
      "post": {
        "tags": [
          "crate::routes::query"
        ],
        "operationId": "query_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Execute SQL query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Query execution error"
          }
        }
      }
    },
    "/api/connections/{id}/query/analyze": {
      "post": {
        "tags": [
          "crate::routes::query"
        ],
        "operationId": "analyze_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Analyze SQL safety",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyReport"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalyzeBody": {
        "type": "object",
        "required": [
          "sql"
        ],
        "properties": {
          "sql": {
            "type": "string"
          }
        }
      },
...
      "ClearHistoryResponse": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ColumnMeta": {
        "type": "object",
        "required": [
          "name",
          "dataTypeID",
          "typeName"
        ],
        "properties": {
          "dataTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "sourceColumn": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "sourceTable": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "typeName": {
            "type": "string"
          }
        }
      },
...
      "HistoryEntry": {
        "type": "object",
        "required": [
          "id",
          "connectionId",
          "sql",
          "startedAt",
          "status"
        ],
        "properties": {
          "connectionId": {
            "type": "string"
          },
          "durationMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "rowCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "sql": {
            "type": "string"
          },
          "startedAt": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "QueryBody": {
        "type": "object",
        "required": [
          "sql"
        ],
        "properties": {
          "confirmDangerous": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "offset": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "sql": {
            "type": "string"
          }
        }
      },
      "QueryResponse": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "requiresConfirmation",
              "report"
            ],
            "properties": {
              "report": {
                "$ref": "#/components/schemas/SafetyReport"
              },
              "requiresConfirmation": {
                "type": "boolean"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "requiresConfirmation",
              "results"
            ],
            "properties": {
              "requiresConfirmation": {
                "type": "boolean"
              },
              "results": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StatementResult"
                }
              }
            }
          }
        ]
      },
...
      "SafetyReport": {
        "type": "object",
        "required": [
          "statements",
          "dangerous",
          "reasons"
        ],
        "properties": {
          "dangerous": {
            "type": "boolean"
          },
          "parseError": {
            "type": [
              "string",
              "null"
            ]
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "statements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatementRisk"
            }
          }
        }
      },
...
      "StatementResult": {
        "type": "object",
        "required": [
          "index",
          "sql",
          "columns",
          "rows",
          "rowCount",
          "durationMs",
          "paginated",
          "limit",
          "offset"
        ],
        "properties": {
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColumnMeta"
            }
          },
          "command": {
            "type": [
              "string",
              "null"
            ]
          },
          "durationMs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "offset": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "paginated": {
            "type": "boolean"
          },
          "rowCount": {
            "type": "integer",
            "minimum": 0
          },
          "rows": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          },
          "source": {
            "type": [
              "object",
              "null"
            ]
          },
          "sql": {
            "type": "string"
          }
        }
      },
      "StatementRisk": {
        "type": "object",
        "required": [
          "command",
          "dangerous",
          "reasons"
        ],
        "properties": {
          "command": {
            "type": "string"
          },
          "dangerous": {
            "type": "boolean"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}